+++ src/tools/catch_submit.js
@@ -109,6 +109,7 @@ function check(button) {
 
   var valid = true;
   var first_invalid;
+  button.form.__hf_form_validation = true;
   get_validated_elements(button.form).map(element => {
     if (! reportValidity(element)) {
       valid = false;
@@ -117,6 +118,7 @@ function check(button) {
       }
     }
   });
+  delete(button.form.__hf_form_validation);
 
   if (valid) {
     submit_form_via(button);